Add binary sensor
