+++ custom_components/delonghi_primadonna/sensor.py
@@ -1,5 +1,3 @@
-from homeassistant.components.binary_sensor import (BinarySensorDeviceClass,
-                                                    BinarySensorEntity)
 from homeassistant.components.sensor import SensorDeviceClass, SensorEntity
 from homeassistant.config_entries import ConfigEntry
 from homeassistant.core import HomeAssistant
@@ -15,8 +13,6 @@ async def async_setup_entry(
     delongh_device = hass.data[DOMAIN][entry.unique_id]
     async_add_entities([
         DelongiPrimadonnaNozzleSensor(delongh_device, hass),
-        DelongiPrimadonnaDescaleSensor(delongh_device, hass),
-        DelongiPrimadonnaFilterSensor(delongh_device, hass),
         DelongiPrimadonnaStatusSensor(delongh_device, hass),
     ])
     return True
@@ -61,52 +57,4 @@ class DelongiPrimadonnaStatusSensor(DelonghiDeviceEntity, SensorEntity):
     @property
     def icon(self):
         result = 'mdi:thumb-up-outline'
-        return result
-
-
-class DelongiPrimadonnaDescaleSensor(DelonghiDeviceEntity, BinarySensorEntity):
-    """
-    Shows if the device needs descaling
-    """
-    _attr_device_class = BinarySensorDeviceClass.PROBLEM
-    _attr_name = 'Descaling'
-
-    @property
-    def native_value(self):
-        return self.device.service
-
-    @property
-    def is_on(self) -> bool:
-        return bool((self.device.service >> 3) % 2)
-
-    @property
-    def icon(self):
-        result = 'mdi:dishwasher'
-        if self.is_on():
-            result = 'mdi:dishwasher-alert'
-        return result
-
-
-class DelongiPrimadonnaFilterSensor(
-        DelonghiDeviceEntity,
-        BinarySensorEntity):
-    """
-    Shows if the filter need to be changed
-    """
-    _attr_device_class = BinarySensorDeviceClass.PROBLEM
-    _attr_name = 'Filter'
-
-    @property
-    def native_value(self):
-        return self.device.service
-
-    @property
-    def is_on(self) -> bool:
-        return bool((self.device.service >> 4) % 2)
-
-    @property
-    def icon(self):
-        result = 'mdi:filter'
-        if self.is_on():
-            result = 'mdi:filter-off'
         return result
